Remove redundant casts
git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@27368 bd178f5f-7bea-4014-844f-495ca19b8805

+++ src/org/starexec/data/database/Benchmarks.java
@@ -801,7 +801,7 @@ public class Benchmarks {
 			newBenchmark.setPath(uniqueDir.getAbsolutePath() + File.separator + benchmarkFile.getName());
 
 			FileUtils.copyFileToDirectory(benchmarkFile, uniqueDir);
-			int benchId = Benchmarks.addAndAssociate(newBenchmark, spaceId, (Integer) null);
+			int benchId = Benchmarks.addAndAssociate(newBenchmark, spaceId, null);
 			if (benchId < 0) {
 				log.error("Benchmark being copied could not be successfully added to the database");
 				return benchId;
